% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{baseFEDRContainer-class}
\alias{baseFEDRContainer-class}
\alias{.baseFEDRContainer}
\title{The r-base Foreach Redis container}
\description{
The r-base Foreach Redis container. See \code{?baseFEDRServerContainer} and
\code{?baseFEDRWorkerContainer}
}

