% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFEDRContainer.R
\name{baseFEDRContainer-commom-parameters}
\alias{baseFEDRContainer-commom-parameters}
\title{Common baseFEDRContainer parameter}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{environment}{List, the environment variables in the container}
}
\description{
Common baseFEDRContainer parameter
}
