% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFEDRContainer.R
\name{baseFEDRServerContainer}
\alias{baseFEDRServerContainer}
\title{Get the r-base Foreach Redis server container}
\usage{
baseFEDRServerContainer(environment = list(), tag = "latest")
}
\arguments{
\item{environment}{List, the environment variables in the container}
}
\value{
a \code{baseFEDRContainer} object
}
\description{
Get the r-base Foreach Redis server container.
}
\examples{
baseFEDRServerContainer()
}
