% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{class}
\name{BiocFEDRContainer-class}
\alias{BiocFEDRContainer-class}
\alias{.BiocFEDRContainer}
\title{The Bioconductor Foreach Redis container}
\description{
The Bioconductor Foreach Redis container. See \code{?BiocFEDRServerContainer} and
\code{?BiocFEDRWorkerContainer}
}

