% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFEDRContainer.R
\name{BiocFEDRContainer-commom-parameters}
\alias{BiocFEDRContainer-commom-parameters}
\title{Common BiocFEDRContainer parameter}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{environment}{List, the environment variables in the container}
}
\description{
Common BiocFEDRContainer parameter
}
