% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFEDRContainer.R
\name{BiocFEDRServerContainer}
\alias{BiocFEDRServerContainer}
\title{Get the Bioconductor Foreach Redis server container}
\usage{
BiocFEDRServerContainer(
  image = "dockerparallel/parallel-redis-server",
  name = "redisRServerContainer",
  environment = list()
)
}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{environment}{List, the environment variables in the container}
}
\value{
a \code{BiocFEDRContainer} object
}
\description{
Get the Bioconductor Foreach Redis server container.
}
\examples{
BiocFEDRServerContainer()
}
