% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-BiocFEDRContainer.R
\name{BiocFEDRWorkerContainer}
\alias{BiocFEDRWorkerContainer}
\title{Get the Bioconductor Foreach Redis worker container}
\usage{
BiocFEDRWorkerContainer(
  image = "dockerparallel/parallel-redis-worker",
  name = "redisRWorkerContainer",
  RPackages = NULL,
  sysPackages = NULL,
  environment = list(),
  maxWorkerNum = 4L
)
}
\arguments{
\item{image}{Character, the container image}

\item{name}{Character, the optional name of the container}

\item{RPackages}{Character, a vector of R packages that will be installed
by \code{AnVIL::install} before connecting with the server}

\item{sysPackages}{Character, a vector of system packages that will be installed
by \verb{apt-get install} before running the R worker}

\item{environment}{List, the environment variables in the container}

\item{maxWorkerNum}{Integer, the maximum worker number in a container}
}
\value{
a \code{BiocFEDRContainer} object
}
\description{
Get the Bioconductor Foreach Redis worker container.
}
\examples{
BiocFEDRWorkerContainer()
}
