% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-BiocFEDRContainer.R
\name{configServerContainerEnv,baseFEDRContainer-method}
\alias{configServerContainerEnv,baseFEDRContainer-method}
\title{Configure the server container environment,}
\usage{
\S4method{configServerContainerEnv}{baseFEDRContainer}(container, cluster, verbose = FALSE)
}
\arguments{
\item{container}{Reference Container Object. The server container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{baseFEDRContainer} object
}
\description{
Configure the server container environment, only the environment
\code{serverPort}, \code{serverPassword} and \code{sshPubKey} will be set in this method
}
