% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-BiocFEDRContainer.R
\name{configWorkerContainerEnv,baseFEDRContainer-method}
\alias{configWorkerContainerEnv,baseFEDRContainer-method}
\title{Configure the worker container}
\usage{
\S4method{configWorkerContainerEnv}{baseFEDRContainer}(container, cluster, workerNumber, verbose = FALSE)
}
\arguments{
\item{container}{Reference Container Object. The worker container.}

\item{cluster}{S4 \code{DockerCluster} object.}

\item{workerNumber}{Integer. The number of workers in a container.}

\item{verbose}{Integer. The verbose level, default 1.}
}
\value{
A \code{baseFEDRContainer} object
}
\description{
Configure the worker container
}
