% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-BiocFEDRContainer.R
\name{getServerContainer,baseFEDRContainer-method}
\alias{getServerContainer,baseFEDRContainer-method}
\title{Get the r-base foreach Redis container}
\usage{
\S4method{getServerContainer}{baseFEDRContainer}(workerContainer)
}
\arguments{
\item{workerContainer}{The worker container.}
}
\value{
A \code{baseFEDRContainer} server container
}
\description{
Get the r-base foreach Redis container from the worker container
}
