% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptor-BiocFEDRContainer.R
\name{getServerContainer,BiocFEDRContainer-method}
\alias{getServerContainer,BiocFEDRContainer-method}
\title{Get the Bioconductor foreach Redis container}
\usage{
\S4method{getServerContainer}{BiocFEDRContainer}(workerContainer)
}
\arguments{
\item{workerContainer}{The worker container.}
}
\value{
A \code{BiocFEDRContainer} server container
}
\description{
Get the Bioconductor foreach Redis container from the worker container
}
